[chore] Add support for ROS 2 Jazzy
Jazzy is the latest ROS 2 release and the current LTS.

this adds the support in a backwards-compatible way by checking for the
presence of the the new `cv_bridge.hpp` header.

the cleaner solution would be to create a branch for Humble support
(if/when any other commit is needed for humble) and just upgrade the
code to Jazzy in a non-backwards-compatible way. however, i chose this
approach since i wasn't sure which you prefer and since there are no
other branches to support other ROS releases and Humble is still a
supported ROS release.